+++ src/net/craftlin/api/value/Effect.kt
@@ -6,7 +6,7 @@ package net.craftlin.api.value
 open class Effect(
     var type: String,
     var time: Long,
-    var strength: Int = 1,
+    var strength: Long = 1,
     var ambient: Boolean = false,
     var particles: Boolean = true,
     var icon: Boolean = true
