+++ src/net/craftlin/bukkit/impl/entity/base/BukkitLivingEntity.kt
@@ -9,8 +9,8 @@ import org.bukkit.attribute.Attribute
 import org.bukkit.potion.PotionEffect
 
 abstract class BukkitLivingEntity(private val livingEntity: org.bukkit.entity.LivingEntity): BukkitEntity(livingEntity), LivingEntity {
-    override var health: Int
-        get() = livingEntity.health.toInt()
+    override var health: Long
+        get() = livingEntity.health.toLong()
         set(value) { livingEntity.health = value.toDouble() }
     override val ignition: Ignition
         get() = Ignition(livingEntity.fireTicks.toLong())
@@ -28,17 +28,17 @@ abstract class BukkitLivingEntity(private val livingEntity: org.bukkit.entity.Li
 
     override val effects: Collection<Effect>
         get() = livingEntity.activePotionEffects.map {
-            Effect(BukkitEffectType.Converter(it.type), it.duration.toLong(), it.amplifier, it.isAmbient, it.hasParticles(), it.hasIcon())
+            Effect(BukkitEffectType.Converter(it.type), it.duration.toLong(), it.amplifier.toLong(), it.isAmbient, it.hasParticles(), it.hasIcon())
         }
 
     override fun effect(effect: Effect) {
         livingEntity.addPotionEffect(with(effect) {
-            PotionEffect(BukkitEffectType.Converter(type), time.toInt(), strength, ambient, particles, icon)
+            PotionEffect(BukkitEffectType.Converter(type), time.toInt(), strength.toInt(), ambient, particles, icon)
         })
     }
 
     override val speed = BukkitEntitySpeed(entity)
     override val maxHealth = BukkitAttribute(entity, Attribute.GENERIC_MAX_HEALTH)
 
-    override fun damage(amount: Int) { livingEntity.damage(amount.toDouble()) }
+    override fun damage(amount: Long) { livingEntity.damage(amount.toDouble()) }
 }
